+++ src/App.vue
@@ -2,6 +2,7 @@
 import { computed } from 'vue'
 import { RouterView, useRoute } from 'vue-router'
 import AppShell from './components/AppShell.vue'
+import LoadingBar from './components/LoadingBar.vue'
 
 const route = useRoute()
 // Routes opting out of the shell (Connect, Picker) set `meta.bare`.
@@ -9,6 +10,7 @@ const showShell = computed(() => route.meta.bare !== true)
 </script>
 
 <template>
+  <LoadingBar />
   <AppShell v-if="showShell">
     <RouterView />
   </AppShell>